Moved deck in progress initialization and fixed typo
