+++ next/lib/deck-utils.js
@@ -1,0 +1,12 @@
+export const initializeDeckBuilder = () => {
+  return {
+    deckName: '',
+    deckPath: '',
+    deckPower: '',
+    deckCoverart: '',
+    mainDeck: [],
+    sideboard: [],
+    errors: [],
+    asText: ''
+  };
+};
